Make protocol_message arbitrarily large
protocol_message had a fixed length of 240000. If a message was larger,
it was just truncated. At the same time, the buffer was fixed to
240000B, which is a very large allocation on the stack.

This commit makes a default stack allocation of 24000B. If a message is
larger, it will be dynamically allocated, similarly to what is done in
tagged_message.

To prevent multiple frees, we disable the copy constructor and only
specify a move constructor.